Update patch set 2
Patch Set 2:

(22 comments)

My main concern is about how we can integrate that into the existing kni-intaller binary, Makefile, etc... there is some code that is mostly a duplicate, and some other that can be integrated into the installer with some work. What is your proposal for the integration?

Patch-set: 2
Label: Verified=0