Add kubevirt deployment in base templates
We enable the deployment of kubevirt as a default for the PAE
blueprint. In order to do it, new SCC, NameSpace, Operator and CR needs
to be created.
It is also needed to modify the kustomize flags, to force to ignore
the native kustomize reorder, that was causing OpenShift resources
to be appended at the end.
Added some modifications to the kustomize workflow as well, to
check for empty resources, and to retry failed applies.

Change-Id: I0dbc561c364fd2eaa44549d47c2e167b007301bf
Signed-off-by: Yolanda Robla <yroblamo@redhat.com>