Adding requirements per profile
Make requirements more flexible, depending on profile,
and also allowing independent versions of kubectl and oc.
Also also need to bump the server versions to 4.1.3,
and the client version to 1.15, because there is
a problem with strategic merge that blocks us when
applying scc. This causes us to have independent versions
for oc and kubectl.

Signed-off-by: Yolanda Robla <yroblamo@redhat.com>

Change-Id: Ib8d03f0a71c1dfe299a4bffe2e7b5f4c8d9b56b8