Update references to point to absolute git
To support sites that are not part of the blueprint, the
workflow has been changed and it first clones the specific
site directory, without cloning the full blueprint. Then, a
reference to the profile needs to exist on each site in order
to recreate the blueprint.
So we need to modify this reference on the sample sites, in
order to provide a complete reference to the workflow.

Signed-off-by: Yolanda Robla <yroblamo@redhat.com>
Change-Id: Ib0691a0ff9ef72590d257e7035907bcd846c4570