Kustomize replicas for ingress router
For some sites and blueprints (0/1 workers), ingress
controller is failing because it has a default of 2.
Create a default manifests so we are able to patch,
and set replicas to 3, that matches what we have
in base for the number of workers. Set it to 1 in the
case of libvirt. This can also be overriden by site

Change-Id: Iaee5c1e3aeccaecdaf21462aacd94a8eab1220b9