WIP: Do tuning manually for realtime nodes
As we do not have support for tuned profile in current
tuning operator, the suggested solution is to disable
the tuned pod in the realtime workers. The steps to achieve
it is unmanage the deployment of the operator, then set
replicas to 0, to stop it managing the daemonset. Then we
can add the affinity to the daemonset to avoid that
realtime workers.
After that, we can add tuning properly with machineconfig,
and this will not be overwritten by the tuning operator.
This is a temporary change until valid realtime support
is added to the tuning operator.

Signed-off-by: Yolanda Robla <yroblamo@redhat.com>
Change-Id: Ic2a8679e0a1015e1a3d9b503dabdf9702761e836